+++ src/algos/statistics.h
@@ -18,7 +18,7 @@
 imstats* statistics(sequence *seq, int image_index, fits *fit, int layer,
 		rectangle *selection, int option, gboolean multithread);
 
-int compute_means_from_flat_cfa(fits *fit, double mean[4]);
+int compute_means_from_flat_cfa(fits *fit, float mean[4]);
 
 void allocate_stats(imstats **stat);
 imstats* free_stats(imstats *stat);
