+++ src/algos/statistics_float.c
@@ -325,7 +325,7 @@ imstats* statistics_internal_float(fits *fit, int layer, rectangle *selection, i
 	return stat;
 }
 
-int compute_means_from_flat_cfa_float(fits *fit, double mean[4]) {
+int compute_means_from_flat_cfa_float(fits *fit, float mean[4]) {
 	int row, col, c, i = 0;
 	float *data;
 	unsigned int width, height;
@@ -352,16 +352,16 @@ int compute_means_from_flat_cfa_float(fits *fit, double mean[4]) {
 	/* Compute mean of each channel */
 	for (row = starty; row < height - 1 - starty; row += 2) {
 		for (col = startx; col < width - 1 - startx; col += 2) {
-			mean[0] += (double) data[col + row * width];
-			mean[1] += (double) data[1 + col + row * width];
-			mean[2] += (double) data[col + (1 + row) * width];
-			mean[3] += (double) data[1 + col + (1 + row) * width];
+			mean[0] += data[col + row * width];
+			mean[1] += data[1 + col + row * width];
+			mean[2] += data[col + (1 + row) * width];
+			mean[3] += data[1 + col + (1 + row) * width];
 			i++;
 		}
 	}
 
 	for (c = 0; c < 4; c++) {
-		mean[c] /= (double) i;
+		mean[c] /= (float) i;
 	}
 	return 0;
 }
