+++ src/algos/statistics_float.h
@@ -5,5 +5,5 @@
 
 imstats* statistics_internal_float(fits *fit, int layer, rectangle *selection,
 		int option, imstats *stats, gboolean multithread);
-int compute_means_from_flat_cfa_float(fits *fit, double mean[4]);
+int compute_means_from_flat_cfa_float(fits *fit, float mean[4]);
 #endif
